Ensure option builders are only ever run once
We ran into this issue when implementing subfactories and found that
they were being recreated in the `.after` step and causing very strange
issues to show up.

By precomputing the options, we can be sure they'll only ever be run
once and therefore are safe to do complex calculations in.